+++ Mods/S7_Config_de8f15f2-65a2-4ee4-a25f-7a7ab0305a58/Story/RawFiles/Lua/S7_ModInterface.lua
@@ -245,7 +245,7 @@ function S7_DynamicAction(option, switch)
     elseif quickMenuVars.selectedAction == "Confirm" then
         --  Call Stats Configurator
         table.insert(toConfigure, {[quickMenuVars.modName] = Ext.JsonStringify(quickMenuVars.configData)})
-        S7_StatsConfigurator()
+        StatsConfigurator()
         S7_StatsSynchronize()
         toConfigure = {}
         S7_BuildConfigData(quickMenuVars.modName, Ext.JsonStringify(quickMenuVars.configData))